rename and simplify example
